requirements.txt
customtkinter
tkiner
CTkMessagebox
CTkToolTip
numpy
matplotlib
pandas
scipy
pyserial
Pillow
